wildcard moved to internal, test fixed
